[FEAT] Add config toggle, to silence invalid/disallowed command feedback messages for linked commands

+++ Common/src/main/java/com/hypherionmc/sdlink/core/discord/hooks/MinecraftCommandHook.java
@@ -34,8 +34,10 @@ public final class MinecraftCommandHook {
             return;
 
         if (!SDLinkConfig.INSTANCE.linkedCommands.allowedChannels.isEmpty() && SDLinkConfig.INSTANCE.linkedCommands.allowedChannels.stream().noneMatch(c -> c.equals(event.getChannel().getId()))) {
-            event.getMessage().reply(Text.translate("commands.not_allowed")).mentionRepliedUser(false).queue(s -> s.delete().queueAfter(5, TimeUnit.SECONDS));
-            event.getMessage().delete().queueAfter(5, TimeUnit.SECONDS);
+            if (SDLinkConfig.INSTANCE.linkedCommands.outputInvalid) {
+                event.getMessage().reply(Text.translate("commands.not_allowed")).mentionRepliedUser(false).queue(s -> s.delete().queueAfter(5, TimeUnit.SECONDS));
+                event.getMessage().delete().queueAfter(5, TimeUnit.SECONDS);
+            }
             return;
         }
 
@@ -77,10 +79,12 @@ public final class MinecraftCommandHook {
             return;
         }
 
-        event.getMessage().reply(Text.translate("commands.command_not_allowed")).mentionRepliedUser(false).queue(suc -> {
-            suc.delete().queueAfter(5, TimeUnit.SECONDS);
-            event.getMessage().delete().queueAfter(5, TimeUnit.SECONDS);
-        });
+        if (SDLinkConfig.INSTANCE.linkedCommands.outputInvalid) {
+            event.getMessage().reply(Text.translate("commands.command_not_allowed")).mentionRepliedUser(false).queue(suc -> {
+                suc.delete().queueAfter(5, TimeUnit.SECONDS);
+                event.getMessage().delete().queueAfter(5, TimeUnit.SECONDS);
+            });
+        }
     }
 
     private static void executeCommand(String command, int permLevel, MessageReceivedEvent event, SDLinkAccount account) {
